requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
pymupdf==1.23.3
requests==2.31.0
openai==1.14.2
